Add JS app
this was created with `yarn create vite` and then following the prompts
to make a typescript react app.

When running `yarn build`, it will compile the app into the `dist` folder.

This will have an index.html file that we can point our Plug.Static plug
at in our Plug server.