Add wallaby tests
this adds a feature tests as well as adds the proper aliases to the mix.exs
file to bundle/build the js app before running the tests.

Within the last few years, several js bundlers have come out that are super
fast, so it is okay to always bundle before running your tests. this
ensures you never have a stale JS app and are wondering why things aren't
loading.